rename list components that are not list
