+++ apps/admin-ui/src/component/reservation-units/ReservationUnits.tsx
@@ -1,12 +1,12 @@
 import { debounce } from "lodash";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { H1 } from "common/src/common/typography";
+import { H1 } from "common";
+import { HR } from "@/component/Table";
+import { Container } from "@/styles/layout";
 import Filters, { FilterArguments, emptyState } from "./Filters";
 import ReservationUnitsDataReader, { Sort } from "./ReservationUnitsDataLoader";
 import BreadcrumbWrapper from "../BreadcrumbWrapper";
-import { HR } from "../lists/components";
-import { Container } from "../../styles/layout";
 
 const ReservationUnits = (): JSX.Element => {
   const [search, setSearch] = useState<FilterArguments>(emptyState);
