+++ apps/admin-ui/src/component/reservations/ReservationsDataLoader.tsx
@@ -6,9 +6,9 @@ import {
   QueryReservationsArgs,
   ReservationType,
 } from "common/types/gql-types";
+import { More } from "@/component/More";
 import { LIST_PAGE_SIZE } from "../../common/const";
 import { useNotification } from "../../context/NotificationContext";
-import { More } from "../lists/More";
 import Loader from "../Loader";
 import { FilterArguments } from "./Filters";
 import { RESERVATIONS_QUERY } from "./queries";
