+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/[id]/review/ApplicationDataLoader.tsx
@@ -7,7 +7,7 @@ import { filterNonNullable } from "common/src/helpers";
 import { LIST_PAGE_SIZE } from "@/common/const";
 import { useNotification } from "@/context/NotificationContext";
 import Loader from "@/component/Loader";
-import { More } from "@/component/lists/More";
+import { More } from "@/component/More";
 import { useSort } from "@/hooks/useSort";
 import { APPLICATIONS_QUERY } from "./queries";
 import { ApplicationsTable, SORT_KEYS } from "./ApplicationsTable";
