+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/[id]/review/ApplicationEventsTable.tsx
@@ -12,7 +12,7 @@ import { truncate } from "@/helpers";
 import { applicationDetailsUrl } from "@/common/urls";
 import { getApplicantName } from "@/component/applications/util";
 import { formatNumber } from "@/common/util";
-import { CustomTable, ExternalTableLink } from "@/component/lists/components";
+import { CustomTable, ExternalTableLink } from "@/component/Table";
 import { ApplicationEventStatusCell } from "./StatusCell";
 
 const formatters = getFormatters("fi");
